Refactor Dentist delete method to use ResponseEntity. Dentist is working. Patient is not working. Error 500 in postman

+++ src/main/java/com/liliana/controller/DentistController.java
@@ -2,23 +2,23 @@ package com.liliana.controller;
 
 import com.liliana.domain.Dentist;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.liliana.services.DentistService;
 
 import java.util.List;
 
-
 @RestController
+@RequestMapping("/dentists")
 public class DentistController {
-
     @Autowired
     private DentistService dentistService;
 
-
-    @GetMapping("/dentist")
+    @GetMapping("/hi-dentist")
     public String WelcomeDentist(){
-        String mostrar = "Welcome, dear dentist";
-        return mostrar;
+        String hello = "Welcome, dear dentist";
+        return hello;
     }
 
     @GetMapping("/dentists")
@@ -32,16 +32,24 @@ public class DentistController {
     }
 
     @DeleteMapping("/dentist/{dentistId}")
-    private void deleteDentist(@PathVariable("dentistId") Integer dentistId)
-    {
-        dentistService.delete(dentistId);
+    private ResponseEntity deleteDentist(@PathVariable("dentistId") Integer dentistId) {
+        ResponseEntity response = null;
+
+        if(dentistId==null){
+            response = new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }else{
+            dentistService.delete(dentistId);
+            response = new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
+
+        return response;
     }
 
-    @PostMapping("/dentists")
-    private int saveDentist(@RequestBody Dentist dentist)
+    @PostMapping("/dentist")
+    private Dentist saveDentist(@RequestBody Dentist dentist)
     {
         dentistService.saveOrUpdate(dentist);
-        return dentist.getId();
+        return dentist;
     }
 
     @PutMapping("/dentist")
